+++ whatslab/src/components/ListMessenger.js
@@ -36,27 +36,63 @@ height: 8vh;
 margin-top: 2px;
 `
 
+class ListMessenger extends React.Component {
+    state = {
+        listMessenger: [
+            {
+                nomeUsuario: '',
+                messenger: ''
+            },
 
-const ListMessenger = (props) => {
+            {
+                nomeUsuario: '',
+                messenger: ''
+            }
+        ],
 
-    const listaDeComponentes = this.state.listMessenger.map((listMessenger) => {
+        valorInputNomeUsuario: "",
+        valorInputMessenger: "",
+    };
+
+    adicionarMessenger = () => {
+        const newMessenger = {
+            nomeUsuario: this.state.valorInputNomeUsuario,
+            messenger: this.state.valorInputMessenger
+        };
+        const novaMensagem = [...this.state.listMessenger, newMessenger];
+
+        this.setState({ listMessenger: novaMensagem })
+    };
+
+    onChangeInputUsuario = (event) => {
+        this.setState({ valorInputNomeUsuario: event.target.value });
+    };
+
+    onChangeInputMensagem = (event) => {
+        this.setState({ valorInputMessenger: event.target.value });
+    };
+
+    render() {
+        const listaDeComponentes = this.state.listMessenger.map((listMessenger) => {
+            return (
+                <p>
+                    {listMessenger.nomeUsuario} - {listMessenger.messenger}
+                </p>
+            )
+        })
+
         return (
-            <p>
-                {listMessenger.nomeUsuario} - {listMessenger.messenger}
-            </p>
+            <SessaoEnviodeMensangem>
+                <InputUsuario value={this.state.valorInputNomeUsuario} onChange={this.onChangeInputUsuario} placeholder="Usuario"></InputUsuario>
+                <InputMensagem value={this.state.valorInputMessenger} onChange={this.onChangeInputMensagem} placeholder="Mensagem"></InputMensagem>
+                <Button onClick={this.adicionarMessenger}>ENVIAR</Button>
+                <div>{listaDeComponentes}</div>
+            </SessaoEnviodeMensangem>
         )
-    })
-
-    return (
-
-        <SessaoEnviodeMensangem>
-            <InputUsuario value={props.inputNomeUsuario} onChange={this.onChangeInputUsuario} placeholder="Usuario"></InputUsuario>
-            <InputMensagem value={this.state.valorInputMessenger} onChange={this.onChangeInputMensagem} placeholder="Mensagem"></InputMensagem>
-            <Button onClick={this.adicionarMessenger}>ENVIAR</Button>
-            <p>{listaDeComponentes}</p>
-        </SessaoEnviodeMensangem>
-    )
+    };
 };
+
+
 
 
 export default ListMessenger;
